change multimodal_object_recognition_ros.cpp --> multimodal_object_recognition_node.cpp
